Chord Part I - Linear Time Search
Correctness is verified using test_part_1.py.
Remember to launch the test soon after the Chord servers start, so that
during test delay the only stabilization won't miss, to keep the result
correct.

Note: The busy loop in chord.cc:main is changed to a condition variable,
but since chord.cc is not allowed to modify, we need to revert this before
release.