Chord Part I - Fix Finger Table Lookup
Finger table lookup was wrongly implemented, resulting in bad
performance. The correct design is to calculate a step size such that
the last entry of the finger table can index to the entire ID space.

Another minor fix in closest_preceding_node() is to return the successor when
finger table is empty.

Correctness is verified by test_part_1.py.
Message complexity is verified by test_part_2.py.

- idle RPC/s = 36.3 < 64
- RPC/find = 2.98 < 7
Passing the baseline.